requirements.txt
asgiref==3.11.1
certifi==2026.4.22
charset-normalizer==3.4.7
coreapi==2.3.3
coreschema==0.0.4
Django==6.0.5
django-filter==23.5
djangorestframework==3.15.1
idna==3.13
itypes==1.2.0
Jinja2==3.1.6
Markdown==3.10.2
MarkupSafe==3.0.3
pillow==12.2.0
python-decouple==3.8
pytz==2026.2
requests==2.33.1
sqlparse==0.5.5
uritemplate==4.2.0
urllib3==2.7.0
